requirements.txt
streamlit>=1.38.0
python-dotenv
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.1.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
